backend/requirements.txt
fastapi
uvicorn[standard]
httpx
cryptography
databases[postgresql]
psycopg2-binary
python-dotenv
PyJWT
rq
redis
pypdf
beautifulsoup4
pgvector
google-genai
bcrypt 
pydantic[email]
python-jose
databases[postgresql]
asyncpg 
psycopg 
instagrapi 
google-api-python-client 
google-auth google-auth-oauthlib 
apscheduler